Custom admin compatibility.
The filebrowser site had hardcoded references to Django's default
admin site django.contrib.admin.sites.site, so it crashed when using
a custom admin site (because it would register itself on the wrong
place and try to access the wrong apps, e.g. for auth).
The filebrowser site now accepts an additional argument to indicate
the appropriate admin site.